requirements.txt
albumentations==1.4.2
comet_ml==3.39.2
coremltools==7.1
Flask==3.0.2
GitPython==3.1.41
GitPython==3.1.42
ipython==8.12.3
matplotlib==3.4.3
mss==9.0.1
nncf==2.9.0
numpy==1.24.4
onnx==1.16.0
onnxruntime==1.17.1
onnxsim==0.4.36
opencv_python==4.9.0.80
openvino==2024.0.0
paddle==1.0.2
pafy==0.5.5
pandas==2.0.3
Pillow==10.2.0
psutil==5.9.8
pycocotools==2.0.7
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.0
tensorflow==2.16.1
tensorflowjs==4.17.0
tensorrt==8.6.1.post1
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
tritonclient==2.44.0
ultralytics==8.1.12
x2paddle==1.4.1
